recommend-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0
implicit==0.7.2
python-dotenv==1.0.0
gunicorn
